[Refactor] - Move glide request builder extensions into image module
